feat(basic): ui 컴포넌트 분리
- AdminPage props 상태값 나누는 중
